Handle only ClojureScript source files
This is a response to the reported bug:

https://twitter.com/esten/status/791389213870661633

I've added both "Exported UTIs" & "Imported UTIs".

As per http://stackoverflow.com/a/7489137/1363227

* Exported UTIs are the UTIs for which your application is
authoritative. That is, document types which you own and define.

* Imported UTIs are document types which you know about but for which
some other application may be authoritative.

Should we leave both, as I don't know of any other application that is
authoritative for ClojureScript on iOS?